+++ src/main/java/com/codetaylor/mc/pyrotech/modules/core/init/ItemInitializer.java
@@ -47,6 +47,8 @@ public final class ItemInitializer {
 
     registry.registerItem(new ItemGloamberrySeeds(), ItemGloamberrySeeds.NAME);
     registry.registerItem(new ItemGloamberries(), ItemGloamberries.NAME);
+
+    registry.registerItem(new ItemFurnaceCore(), ItemFurnaceCore.NAME);
   }
 
   @SideOnly(Side.CLIENT)
@@ -88,7 +90,9 @@ public final class ItemInitializer {
           ModuleCore.Items.GLOAMBERRIES,
           ModuleCore.Items.GLOAMBERRY_WINE,
 
-          ModuleCore.Items.TAINTED_MEAT
+          ModuleCore.Items.TAINTED_MEAT,
+
+          ModuleCore.Items.FURNACE_CORE
       );
 
       // Material