+++ src/main/java/com/codetaylor/mc/pyrotech/modules/core/item/ItemFurnaceCore.java
@@ -1,0 +1,10 @@
+package com.codetaylor.mc.pyrotech.modules.core.item;
+
+import net.minecraft.item.Item;
+
+public class ItemFurnaceCore
+    extends Item {
+
+  public static final String NAME = "furnace_core";
+
+}